Add files via upload
Following are the changes:
1) arrange_cross_features.m
This code is used to arrange cross feature in class wise manner. Original features are e.g. 1-1, 2-2, 3-3, 1-2, 1-3, 2-1,2-3, 3-1,3-2.
They are arranged into following form: 1-1,1-2,1-3,2-1,2-2,2-2,3-1,3-2,3-3
This is done to ease the design of feature fusion classifier.
2) feature_fusion_classifier.py
This contains keras code for the feature fusion classifier. Cross features are obtained from aec. For apy 1,2,4,6 (~85%), for apy 8, 12, 17, 23, 24(~80%).
3) cnn_features_classifier.py
This contain code for original 4096 dim cnn feature classifier using the same network for cross features. For apy 8,12,17,23,24 (~50 to 80)
4) mnist_keras.py
sample keras code for mnist classification